backend/requirements.txt
fastapi
uvicorn
httpx
websockets
pytest
pytest-asyncio
pytest-cov
PyYAML
prisma
python-jose
passlib[bcrypt]
asgi-lifespan
python-dotenv
python-multipart
cryptography
tenacity
fastapi-sso
redis
